requirements.txt
aiohttp==3.6.3
async-timeout==3.0.1
attrs==20.3.0
bpemb==0.3.2
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
cloudpickle==1.6.0
cycler==0.10.0
dataclasses==0.6
decorator==4.4.2
Deprecated==1.2.10
discord.py==1.5.1
filelock==3.0.12
flair==0.6.1.post1
ftfy==5.8
future==0.18.2
gdown==3.12.2
gensim==3.8.3
hyperopt==0.2.5
idna==2.10
Janome==0.4.1
joblib==0.17.0
kiwisolver==1.3.1
konoha==4.6.2
langdetect==1.0.8
lxml==4.6.2
matplotlib==3.3.3
mpld3==0.3
multidict==4.7.6
networkx==2.5
numpy==1.19.4
overrides==3.0.0
packaging==20.7
Pillow==8.0.1
protobuf==3.14.0
pyparsing==2.4.7
python-dateutil==2.8.1
regex==2020.11.13
requests==2.25.0
sacremoses==0.0.43
scikit-learn==0.23.2
scipy==1.5.4
segtok==1.5.10
sentencepiece==0.1.94
six==1.15.0
smart-open==4.0.1
sqlitedict==1.7.0
tabulate==0.8.7
threadpoolctl==2.1.0
tokenizers==0.9.3
torch==1.7.0
tqdm==4.54.0
transformers==3.5.1
typing-extensions==3.7.4.3
urllib3==1.26.2
wcwidth==0.2.5
wrapt==1.12.1
yarl==1.5.1
